Issue #421 Move temporal extent functions to separate module
